funRegulationBackCode/requirements.txt
amqp==2.4.2
bcbio-gff==0.6.6
billiard==3.6.0.0
biopython==1.74
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.1.7
django-celery-results==1.0.4
idna==2.8
kombu==4.5.0
ncbi-datasets-pylib==12.13.2
numpy==1.17.2
psycopg2-binary==2.8.6
pytz==2018.9
requests==2.22.0
six==1.12.0
urllib3==1.25.6
vine==1.3.0
xmltramp2==3.1.1
yte==1.5.1
zipp==3.11.0